\name{make.gene.densities.plot}
\alias{make.gene.densities.plot}
\title{Make a density plot for all genes}
\description{Make a density plot for all genes}
\usage{
make.gene.densities.plot(cnas, fname.stem, xlab)
}
\arguments{
  \item{cnas}{A matrix with CNA info, genes X samples}
  \item{fname.stem}{To use in filename, description of how CNAs were called is suggested}
  \item{xlab}{A description of type of CNA data}
}
\details{Make a density plot for all genes}
\value{None}
\author{Cindy Yao and Emilie Lalonde}
